Still working on Makefile

+++ src/data_type.h
@@ -1,5 +1,5 @@
 struct process{
 	char * name;
-	int pid;
+	unsigned long pid;
 };
